Do not indent parentheses inside string
Closes #7.

This was tricky. In commit 6c2609dc77a0a092d0c9c7f266238922a160171f,
we made lexer always go to the beginning/end of the string, when
invoked inside a string. SMIE, however, to some extent cares about
parentheses inside string so it caused problems. SMIE cares about
parens in strings because `smie-indent-close', which tries to align
closing paren, is called before `smie-indent-inside-string', which
disables indentation in strings. `smie-indent-close' tries to find
matching paren and due to the mentioned commit, it thought that
matching "paren" is at the beginning of the string and indented the
closing paren to that.

We solve this by lexer looking for parens before looking for strings.
In addition to that the logic of lexer is reworked so that comments
and strings are never matched by meson-token-spec, but always based
on (syntax-ppss). Hopefully, this will make the lexer hopefully.