Use display-buffer instead of switch-to-buffer to show documentation
Using switch-to-buffer to show the manual is not ideal because it
always shows the Meson manual in the currently selected window. If the
manual is already shown in another window, the user ends up with two
windows showing the manual. This commit prevents this by using
display-buffer with appropriate actions so that the manual is shown in
the same window where it was before. Only if the manual is not
visible, it opens in the selected window.

Another advantage of display-buffer is that it is user-configurable.
To that end, this commits introduces meson-doc-display-buffer-action,
which can be customized if someone prefers to select windows for
displaying the manual differently.